chore: #173 Move CustomTextFieldRepresentable to PageInputTextField
